+++ temba/contacts/models.py
@@ -757,7 +757,7 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
             Q(contacts__in=[self]) | Q(urns__in=contact_urns) | Q(groups__in=contact_groups)
         )
 
-        return scheduled_broadcasts.order_by("schedule__next_fire")
+        return scheduled_broadcasts.select_related("org").order_by("schedule__next_fire")
 
     def get_history(self, after: datetime, before: datetime, include_event_types: set, limit: int) -> list:
         """
@@ -815,8 +815,8 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
         )
 
         ticket_events = (
-            TicketEvent.objects.filter(created_on__gte=after, created_on__lt=before, ticket__contact=self)
-            .select_related("ticket")
+            TicketEvent.objects.filter(created_on__gte=after, created_on__lt=before, contact=self)
+            .select_related("ticket__ticketer")
             .order_by("-created_on")[:limit]
         )