+++ temba/tickets/models.py
@@ -166,31 +166,41 @@ class Ticket(models.Model):
     # the status of this ticket, one of open, closed, expired
     status = models.CharField(max_length=1, choices=STATUS_CHOICES)
 
-    # when this ticket was opened
+    # when this ticket was opened, closed, modified
     opened_on = models.DateTimeField(default=timezone.now)
-
-    # when this ticket was last modified
+    closed_on = models.DateTimeField(null=True)
     modified_on = models.DateTimeField(default=timezone.now)
 
-    # when this ticket was closed
-    closed_on = models.DateTimeField(null=True)
+    # when this ticket last had activity which includes messages being sent and received, and is used for ordering
+    last_activity_on = models.DateTimeField(default=timezone.now)
 
     assignee = models.ForeignKey(User, on_delete=models.PROTECT, null=True, related_name="assigned_tickets")
 
     def assign(self, user: User, *, assignee: User, note: str):
+        now = timezone.now()
         self.assignee = assignee
-        self.modified_on = timezone.now()
-        self.save(update_fields=("assignee", "modified_on"))
+        self.modified_on = now
+        self.last_activity_on = now
+        self.save(update_fields=("assignee", "modified_on", "last_activity_on"))
 
         self.events.create(
-            org=self.org, event_type=TicketEvent.TYPE_ASSIGNED, assignee=assignee, note=note, created_by=user
+            org=self.org,
+            contact=self.contact,
+            event_type=TicketEvent.TYPE_ASSIGNED,
+            assignee=assignee,
+            note=note,
+            created_by=user,
         )
 
     def add_note(self, user: User, *, note: str):
+        now = timezone.now()
         self.modified_on = timezone.now()
-        self.save(update_fields=("modified_on",))
+        self.last_activity_on = now
+        self.save(update_fields=("modified_on", "last_activity_on"))
 
-        self.events.create(org=self.org, event_type=TicketEvent.TYPE_NOTE, note=note, created_by=user)
+        self.events.create(
+            org=self.org, contact=self.contact, event_type=TicketEvent.TYPE_NOTE, note=note, created_by=user
+        )
 
     @classmethod
     def bulk_close(cls, org, user, tickets):
@@ -216,9 +226,11 @@ class Ticket(models.Model):
     class Meta:
         indexes = [
             # used by the open tickets view
-            models.Index(name="tickets_org_open", fields=["org", "-opened_on"], condition=Q(status="O")),
+            models.Index(name="tickets_org_open", fields=["org", "-last_activity_on", "-id"], condition=Q(status="O")),
             # used by the closed tickets view
-            models.Index(name="tickets_org_closed", fields=["org", "-opened_on"], condition=Q(status="C")),
+            models.Index(
+                name="tickets_org_closed", fields=["org", "-last_activity_on", "-id"], condition=Q(status="C")
+            ),
             # used by the tickets filtered by ticketer view
             models.Index(name="tickets_org_ticketer", fields=["ticketer", "-opened_on"]),
             # used by the list of tickets on contact page and also message handling to find open tickets for contact
@@ -248,6 +260,7 @@ class TicketEvent(models.Model):
 
     org = models.ForeignKey(Org, on_delete=models.PROTECT, related_name="ticket_events")
     ticket = models.ForeignKey(Ticket, on_delete=models.PROTECT, related_name="events")
+    contact = models.ForeignKey(Contact, on_delete=models.PROTECT, related_name="ticket_events")
     event_type = models.CharField(max_length=1, choices=TYPE_CHOICES)
     note = models.TextField(null=True)
     assignee = models.ForeignKey(User, on_delete=models.PROTECT, null=True, related_name="ticket_assignee_events")
@@ -256,3 +269,9 @@ class TicketEvent(models.Model):
         settings.AUTH_USER_MODEL, on_delete=models.PROTECT, null=True, related_name="ticket_events"
     )
     created_on = models.DateTimeField(default=timezone.now)
+
+    class Meta:
+        indexes = [
+            # used for contact history
+            models.Index(name="ticketevents_contact_created", fields=["contact", "created_on"])
+        ]